pyproject.toml
[project]
name = "df_metadata_customizer"
version = "1.2.1"
description = "Open-source MP3 metadata editor built with CustomTkinter, specifically designed for managing fan cover song collections."
readme = "README.md"
requires-python = ">=3.10, <3.15"
dependencies = [
    "customtkinter==5.2.2",
    "mutagen==1.47.0",
    "pillow==12.0.0",
    "polars==1.36.1",
    "rich==14.2.0",
    "tinytag==2.2.0",
]

[dependency-groups]
dev = [
    "pyinstaller==6.17.0",
    "ruff==0.14.10",
]

[tool.setuptools.packages.find]
include = ["df_metadata_customizer"]

[tool.ruff]
line-length = 120
target-version = "py310"
select = [
    "ALL",
]
ignore = [
    "PLR0915", # too-many-statements
    "C901",    # too-complex
    "PLR0911", # too-many-return-statements
    "PLR0912", # too-many-branches
]
